backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
PyMuPDF==1.24.5
reportlab==4.2.0
httpx==0.27.0
PyYAML==6.0.1
python-dotenv==1.0.1
sqlmodel==0.0.19
psycopg2-binary==2.9.9
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
ultralytics==8.2.2
huggingface_hub==0.23.0
surya-ocr==0.4.14
pdf2zh>=1.9.0
